Update employeeLoading.js
randomuser api is pulling the data that I need now. had to decide between pulling specific details onyl and being limited to one record at a time or pulling the number of records I wanted and then picking out what I needed from those records. Opted for the second. Easy to use dot notation to grab the appropriate on as needed this way.